azure_function/requirements.txt
# Azure Functions dependencies
azure-functions==1.17.0
azure-storage-blob==12.19.0

# ML dependencies for ALS model
numpy==1.24.3
scipy==1.11.0
implicit==0.7.2